refactor load_dataset and add mask_variables
